Refact(dashboard): separação de responsabilidades entre controller e view
