+++ dashboard/index.html
@@ -26,7 +26,7 @@
             <button type="button" class="toggle-switch" id="toggleButton">
                 <div class="toggle-slider"></div>
             </button>
-            <span class="toggle-label" id="toggleLabel">Desativado (0)</span>
+            <span class="toggle-label" id="toggleLabel"></span>
         </div>
     </div>
